+++ Changelog.md
@@ -4,3 +4,4 @@ Version BuildSys Test
 * Trivial files are now built
 * Functions for foods are now built
 * Builds on both Windows and Unix-like systems
+* ver_check is now built